+++ apps/rosen/src/app/(main)/events/[details]/Process.tsx
@@ -10,6 +10,7 @@ import {
   type EventProcessesProps,
   formatDateTime,
   type IconProps,
+  Typography,
 } from '@rosen-bridge/ui-kit';
 import { fetcher } from '@rosen-ui/swr-helpers';
 
@@ -352,19 +353,37 @@ export const Process = ({ id, flowId }: { id: string; flowId: string | undefined
 
   const [guardPublicKey, setGuardPublicKey] = useState<string | undefined>();
 
-  const [open, setOpen] = useState(false);
+  const {
+    data: events,
+    error: eventsError,
+    isLoading: eventsIsLoading,
+    mutate: eventMutate,
+  } = useSWR<EventDetailsType[]>(`/v1/events/${id}`, fetcher);
 
-  const url = guardPublicKey
-    ? `/v1/events/${id}/status?triggerTxId=${flowId}&guardPublicKey=${guardPublicKey}`
-    : `/v1/events/${id}/status?triggerTxId=${flowId}`;
-
-  const { data, error, isLoading, mutate } = useSWR<EventStatusType>(
-    open && flowId ? url : null,
+  const {
+    data: status,
+    error: statusError,
+    isLoading: statusIsLoading,
+    mutate: statusMutate,
+  } = useSWR<EventStatusType>(
+    flowId && guardPublicKey
+      ? `/v1/events/${id}/status?triggerTxId=${flowId}&guardPublicKey=${guardPublicKey}`
+      : null,
     fetcher,
   );
 
-  const items = useMemo(() => {
+  const error = eventsError || statusError;
+  const isLoading = eventsIsLoading || statusIsLoading || !flowId;
+  const mutate = eventsError ? eventMutate : statusMutate;
+
+  const data = guardPublicKey ? status : events?.find((event) => event.txId === flowId);
+
+  const isCustomStatus = typeof data?.status === 'object';
+
+  const items = useMemo<EventProcessesProps['items']>(() => {
     if (!data) return [];
+
+    if (typeof data.status === 'object') return [];
 
     const info = structuredClone(pick(steps, findPath(steps, data.status)));
 
@@ -415,22 +434,23 @@ export const Process = ({ id, flowId }: { id: string; flowId: string | undefined
   return (
     <Section
       action={
-        <ProcessSelect disabled={isLoading} value={guardPublicKey} onChange={setGuardPublicKey} />
+        !isCustomStatus && (
+          <ProcessSelect disabled={isLoading} value={guardPublicKey} onChange={setGuardPublicKey} />
+        )
       }
-      collapsible
       error={error}
       load={mutate}
       badge="New"
       title="Progress"
-      onOpenChange={setOpen}
     >
-      <EventProcesses
-        items={items}
-        loading={isLoading}
-        style={{ width: '100%', height: isLoading ? '104px' : undefined }}
-        value={active}
-        onChange={setActive}
-      />
+      {isCustomStatus ? (
+        <Typography>
+          No progress chart is available because this event deviated from the standard lifecycle.
+          The reason for this exception is noted above.
+        </Typography>
+      ) : (
+        <EventProcesses items={items} loading={isLoading} value={active} onChange={setActive} />
+      )}
     </Section>
   );
 };