+++ packages/ui-kit/src/components/cardHeader/component.tsx
@@ -1,8 +1,6 @@
 import { useConfig } from '@/hooks';
 import type { ElementBaseProps, OverridableType } from '@/types';
 
-import './styles.css';
-
 export interface CardHeaderOverrides {}
 
 export type CardHeaderOwnProps = {};
@@ -14,7 +12,7 @@ export type CardHeaderProps = OverridableType<CardHeaderBaseProps, CardHeaderOve
 export const CardHeader = (props: CardHeaderProps) => {
   const { ...rest } = useConfig('CardHeader', props);
 
-  return <div {...rest} />;
+  return <div data-surface="header" {...rest} />;
 };
 
 CardHeader.displayName = 'CardHeader';