+++ packages/ui-kit/src/components/dialogBody/component.tsx
@@ -1,0 +1,18 @@
+import { useConfig } from '@/hooks';
+import type { ElementBaseProps, OverridableType } from '@/types';
+
+export interface DialogBodyOverrides {}
+
+export type DialogBodyOwnProps = {};
+
+export type DialogBodyBaseProps = ElementBaseProps<'div', DialogBodyOwnProps>;
+
+export type DialogBodyProps = OverridableType<DialogBodyBaseProps, DialogBodyOverrides, never>;
+
+export const DialogBody = (props: DialogBodyProps) => {
+  const { ...rest } = useConfig('DialogBody', props);
+
+  return <div data-surface="body" {...rest} />;
+};
+
+DialogBody.displayName = 'DialogBody';