+++ packages/ui-kit/src/components/dialogFooter/component.tsx
@@ -1,8 +1,6 @@
 import { useConfig } from '@/hooks';
 import type { ElementBaseProps, OverridableType } from '@/types';
 
-import './styles.css';
-
 export interface DialogFooterOverrides {}
 
 export type DialogFooterOwnProps = {};
@@ -18,7 +16,7 @@ export type DialogFooterProps = OverridableType<
 export const DialogFooter = (props: DialogFooterProps) => {
   const { ...rest } = useConfig('DialogFooter', props);
 
-  return <div {...rest} />;
+  return <div data-surface="footer" {...rest} />;
 };
 
 DialogFooter.displayName = 'DialogFooter';