+++ packages/ui-kit/src/components/legacy/AppInfo.tsx
@@ -5,6 +5,15 @@ import { Typography } from '@mui/material';
 
 import type { Network as NetworkType } from '@rosen-ui/types';
 
+import {
+  Dialog,
+  DialogBody,
+  DialogCloseButton,
+  DialogHeader,
+  DialogIcon,
+  DialogTitle,
+} from '@/components';
+
 import { useToast } from '../../hooks';
 import { Avatar } from '../avatar';
 import { Icon, type IconProps } from '../icon';
@@ -12,9 +21,6 @@ import { IconButton } from '../iconButton';
 import { Network } from '../network';
 import { Stack } from '../stack';
 import { Divider } from './Divider';
-import { EnhancedDialog } from './EnhancedDialog';
-import { EnhancedDialogContent } from './EnhancedDialogContent';
-import { EnhancedDialogTitle } from './EnhancedDialogTitle';
 
 type NetworkHeight = {
   network: NetworkType;
@@ -74,13 +80,15 @@ export const AppInfo = ({ children, resolver }: AppInfoProps) => {
 
   return (
     <div>
-      <EnhancedDialog maxWidth="tablet" open={open} stickOn="tablet" onClose={() => setOpen(false)}>
-        <EnhancedDialogTitle icon="ExclamationCircle" onClose={() => setOpen(false)}>
-          <Typography variant="h2" fontWeight="bold">
+      <Dialog open={open} unstick="tablet" width="small" onClose={() => setOpen(false)}>
+        <DialogHeader>
+          <DialogIcon name="ExclamationCircle" />
+          <DialogTitle variant="h2" fontWeight="bold">
             About Rosen Bridge
-          </Typography>
-        </EnhancedDialogTitle>
-        <EnhancedDialogContent>
+          </DialogTitle>
+          <DialogCloseButton />
+        </DialogHeader>
+        <DialogBody>
           <Stack spacing={1}>
             <Divider borderStyle="dashed">
               <Typography noWrap color="text-secondary" variant="body2">
@@ -118,8 +126,8 @@ export const AppInfo = ({ children, resolver }: AppInfoProps) => {
             )}
           </Stack>
           {children}
-        </EnhancedDialogContent>
-      </EnhancedDialog>
+        </DialogBody>
+      </Dialog>
 
       <IconButton color="inherit" disabled={loading} loading={loading} onClick={handleClick}>
         <Icon name="InfoCircle" />