+++ packages/ui-kit/src/hooks/useConfig.ts
@@ -1,27 +1,38 @@
 import { useContext } from 'react';
 
-import { ConfigContext } from '@/Providers';
+import { ConfigContext } from '@/providers';
 
 type WithClassName = {
   className?: string;
+  preset?: string;
 };
 
 export const useConfig = <C extends string, P extends WithClassName>(
   componentName: C,
   props: P,
 ): P => {
-  const config = useContext(ConfigContext);
+  const { className, preset, ...inlineProps } = props;
 
-  const baseClass = `Rosen${componentName}`;
+  const configs = useContext(ConfigContext);
 
-  const className = props.className ? `${baseClass} ${props.className}` : baseClass;
+  /**
+   * TODO: remove the inline Biome comment
+   * local:ergo/rosen-bridge/ui#441
+   */
+  // biome-ignore lint/suspicious/noExplicitAny: Use a better type
+  const config = configs?.components?.[componentName as keyof typeof configs.components] as any;
 
-  const global =
-    config?.components?.[componentName as keyof typeof config.components]?.defaultProps;
+  const classes = className ? `Rosen${componentName} ${className}` : `Rosen${componentName}`;
 
-  return {
-    ...global,
-    ...props,
-    className,
-  };
+  const defaultProps = config?.defaultProps;
+
+  const presetProps = config?.presets?.[preset || defaultProps?.preset];
+
+  const result = { ...defaultProps, ...presetProps, ...inlineProps, className: classes };
+
+  if (result.preset) {
+    result['data-preset'] = result.preset;
+  }
+
+  return result;
 };