+++ networks/firo/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @rosen-network/firo
 
+## 0.1.2
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/icons@4.0.0
+
 ## 0.1.1
 
 ### Patch Changes